+++ gherkin/c/src/dialect.c
@@ -2423,8 +2423,8 @@ static const Keywords ru_examples_keywords = { 1, ru_examples_KEYWORDS };
 static const wchar_t* const ru_feature_KEYWORDS[] = { L"Функция", L"Функциональность", L"Функционал", L"Свойство" };
 static const Keywords ru_feature_keywords = { 4, ru_feature_KEYWORDS };
 
-static const wchar_t* const ru_given_KEYWORDS[] = { L"* ", L"Допустим ", L"Дано ", L"Пусть ", L"Если " };
-static const Keywords ru_given_keywords = { 5, ru_given_KEYWORDS };
+static const wchar_t* const ru_given_KEYWORDS[] = { L"* ", L"Допустим ", L"Дано ", L"Пусть " };
+static const Keywords ru_given_keywords = { 4, ru_given_KEYWORDS };
 
 static const wchar_t* const ru_scenario_KEYWORDS[] = { L"Сценарий" };
 static const Keywords ru_scenario_keywords = { 1, ru_scenario_KEYWORDS };
@@ -2435,8 +2435,8 @@ static const Keywords ru_scenarioOutline_keywords = { 1, ru_scenarioOutline_KEYW
 static const wchar_t* const ru_then_KEYWORDS[] = { L"* ", L"То ", L"Затем ", L"Тогда " };
 static const Keywords ru_then_keywords = { 4, ru_then_KEYWORDS };
 
-static const wchar_t* const ru_when_KEYWORDS[] = { L"* ", L"Когда " };
-static const Keywords ru_when_keywords = { 2, ru_when_KEYWORDS };
+static const wchar_t* const ru_when_KEYWORDS[] = { L"* ", L"Когда ", L"Если " };
+static const Keywords ru_when_keywords = { 3, ru_when_KEYWORDS };
 
 static const Dialect ru_dialect = {
         L"ru",